+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file.
 - Added additional colored Hearthstone items. Innkeeper's will choose a random color to sell.
 _All Hearthstones in a player's inventory, and ender chest, share the same cooldown._
 - Added ability to cook food in Fireplaces. _Fireplaces accept the same recipes as Smokers._
+- Added `enableFireplaceCooking` option.
+- Added `chanceForSmokeVariation` option.
+- Added `hearthstoneSafeTeleportRange` option.
 - Added `hearthstone` item tag.
 - Added `igniters` item tag.
 - Added `extinguishers` item tag.